requirements.txt
google-generativeai
langchain_core
langchain_google_genai